CS699:Fix intended rotten behavior for comestibles
I'm pretty sure the DevTeam didn't mean to check the age of the
comestible against its BUC status.

This bug also exists in vanilla NetHack... it's been there since 3.1.0,
released in 1993, almost 20 years ago!

+++ libnitrohack/src/eat.c
@@ -2017,7 +2017,7 @@ int doeat(struct obj *otmp)	/* generic "eat" command funtion (see cmd.c) */
 	    victual.reqtime = objects[otmp->otyp].oc_delay;
 	    if (otmp->otyp != FORTUNE_COOKIE &&
 		(otmp->cursed ||
-		 (((moves - otmp->age) > (int) otmp->blessed ? 50:30) &&
+		 (((moves - otmp->age) > (otmp->blessed ? 50 : 30)) &&
 		(otmp->orotten || !rn2(7))))) {
 
 		if (rottenfood(otmp)) {
